// src/enums/alerts-message.enum.ts
export enum AlertsMessageEnum {
    GenericServiceError = 'Não foi possível realizar a operação. Por favor, tente novamente mais tarde',
    SuccessRegister = 'Cadastro realizado com sucesso',
    InvalidCredentials = 'Credenciais inválidas',
    InvalidToken = 'Token inválido',
    BlankPassword = 'A senha não pode ficar em branco', 
    BlankName = 'O nome não pode ficar em branco', 
    BlankEmail = 'O email não pode ficar em branco', 
    InvalidEmail = 'Email inválido', 
    EmailTaken = 'Email já cadastrado', 
}